+++ cli/create.h
@@ -9,10 +9,12 @@
 extern "C" {
 #endif
 
-int rawstor_cli_create(const char* target, uint64_t size);
+int rawstor_cli_create(
+    const char* target, uint64_t size, uint32_t mirror_count
+);
 
 int rawstor_cli_create_at(
-    const char* location, const char* uuid, uint64_t size
+    const char* location, const char* uuid, uint64_t size, uint32_t mirror_count
 );
 
 #ifdef __cplusplus